[TASK] Extract QueryBuilder from SearchResultSetService/SuggestService/Repository
This pr:

* Extracts the QueryBuilder from the SearchResultSetService/SuggestService/Repository to centralize the logic
where queries get build

Fixes: #1736
